requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
tokenizers>=0.13.0
opencv-python>=4.7.0
Pillow>=9.5.0
timm>=0.9.0
einops>=0.6.0
pydicom>=2.3.0
nibabel>=5.1.0
nltk>=3.8
rouge-score>=0.1.2
sacrebleu>=2.3.0
sentence-transformers>=2.2.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.10.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
wandb>=0.15.0
tqdm>=4.65.0
pyyaml>=6.0
psutil>=5.9.0
rich>=13.3.0
torchviz>=0.0.2
flash-attn>=2.0.0
deepspeed>=0.9.0
